+++ CHANGELOG.md
@@ -330,8 +330,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - On-the-fly Orange add-on registration.
 - Add-on setup tools.
 
-[unreleased]: https://github.com/ewoks-kit/ewoksorange/compare/v4.0.0rc1...HEAD
-[4.1.0rc1]: https://github.com/ewoks-kit/ewoksorange/compare/v4.0.0...v4.0.0rc1
+[unreleased]: https://github.com/ewoks-kit/ewoksorange/compare/v4.1.0rc1...HEAD
+[4.1.0rc1]: https://github.com/ewoks-kit/ewoksorange/compare/v4.0.0...v4.1.0rc1
 [4.0.0]: https://github.com/ewoks-kit/ewoksorange/compare/v3.3.1...v4.0.0
 [3.3.1]: https://github.com/ewoks-kit/ewoksorange/compare/v3.3.0...v3.3.1
 [3.3.0]: https://github.com/ewoks-kit/ewoksorange/compare/v3.2.1...v3.3.0